Implement db.Dump method
After Watcher connects to the replication slot, it needs to read current
database state to avoid inconsistency between the source database and
Etcd. There is no dedicated mechanism in PostgreSQL to do that. The only thing
that it offers is transaction snapshot ID, that can be used with `SET
TRANSACTION SNAPSHOT` command. Then dumping of the database can be done using
series of SELECT statements inside a transaction with snapshot loaded.

That is the purpose of Dump method. We want to use it to select all data
from all tables with specific transaction snapshot and push it to Etcd.
A similar mechanism can be found in Debezium (https://github.com/debezium/debezium).

Change-Id: I0e7fead4a649bc4c95cb17dc14ba3d8822ad06b8
Partial-Bug: #1758826