(WIP) Fix quota performance issue.
This fixes an issue with quota implementation, where project was
retrieved together with all its children to count the number of
resources of particular type. Now it uses a COUNT query instead.

The change also fixes a bug where we sometimes tried to retrieve
a parent project for resources which didn't have one.

Furthermore it adds a simple benchmark for creation of virtual
networks in the presence of other virtual networks in the system.

Finally some logging code was refactored so that the global
logger is not being configured in multiple places.

Closes-Bug: #1789458
Change-Id: I33c0c3a04f528374ac0bf3896d95e71344469819