[WIP] VirtualNetwork: Allocate VxLAN ID
LogicalRouter: Allocate VxLAN ID for the internal VN
using the owner VirtualNetwork uses,
so that no conflict happens when VirtualNetwork allocates the same ID.
This is necessary because LogicalRouter has to allocate a new ID
when none is specified, and it has no way to pass this request
to VirtualNetwork when creating it.

TODO Update the allocation in Update, Delete.

Change-Id: I93b32f08f4fa64c0a682f6f18ced5bc3e80f55cf
Depends-On: I685b197376c0b0ccf99b2465a3c5890ae68f75a1
Closes-Jira-Bug: JBE-868