[WIP] VirtualNetwork: Allocate VxLAN ID
LogicalRouter: Allocate VxLAN ID for the internal VN
using the owner VirtualNetwork uses, so that no conflict happens
when VirtualNetwork allocates the same ID in the same pool.
Using the same pool is necessary
because LogicalRouter has to allocate a new ID when none is specified,
and it has no way to delegate this to a VirtualNetwork when creating it.

VirtualNetwork tests: Use VxLAN ID = 5 instead of 1
  Now that IDs get allocated, they have to be in the int pool's range.
  The network ID pool starts from 5,
  because the resources from init_data use the lower IDs
  and inserting init_data bypasses network ID allocation.

LogicalRouter: Fix deleting the internal VN.
  Deleting the internal VN failed
  because the LR still had a reference to it.
  As a consequence, LR deletion failed.
  Fix this.

Co-authored-by: "Katarzyna Rybacka <katarzyna.rybacka@codilime.com>"

Change-Id: I93b32f08f4fa64c0a682f6f18ced5bc3e80f55cf
Closes-Jira-Bug: JBE-868