Use right version of ansible for adding new compute
There are dependencies with other packages with the docker ansible modules which
cause issues with certain versions of ansible.
There are also bugs in Ansible itself in the latest versions which prevent us
from using the latest version of Ansible. The latest version of ansible will
work for normal provision, but will not work when we try to add computes.
Refer here: https://bugs.launchpad.net/kolla-ansible/+bug/1784123
So, for the moment, use Ansible 2.5.2 (pip install ansible==2.5.2)

Change-Id: I88d1dd7b553b5d31784a1f948ca3887471134219
Closes-Jira-Bug: CEM-2721