Refactor sync to process events
This was required as first step of transaction handling in sync.

Since Sync creation there were a lot of additions to event procesing
logic. Those features could be used for code simplification.

Removed sink concept in favor of simpler row to event processing.

In next step events are going to be aggergated and then pushed to
processor after a "commit" message.

Change-Id: I518a88e5a11e76fdb321b07e64364a450a96044e