allow authenticated docker registry
On authenticated registry (Nexus), we cannot push without secret.
Thus we disable docker push altogether in run stage of `check`,
as we have no secret available there.

Introduce parent contrail-docker-push-base that holds the secret for
logging to new docker registry and in trusted post-run pushes all
existing `docker images`.

DRY, remove the functionality that is implemented in a new parent.

Change-Id: I94044afc75879d1b3383c7aef539ff7683212a5a