Update patch set 4
Patch Set 4: Code-Review-1

I've talked with Józef and learned that this change does not solve the problem.

Currently, there is one replicator, which starts when the GoAPI server starts and begins replication (dumping the database state and then streaming the changes) immediately. To replicate each event, it sends it to each endpoint from an in-memory list. Endpoints are added to the list via an API at runtime, and each event is sent only to the endpoints registered at the moment.

This means that a newly-registered endpoint will only get events generated after its registration.

To solve this, there should be a separate replicator for each endpoint: when an endpoint is registered, a new replicator should be started, which will dump the database and stream the changes to that endpoint. Then the endpoint will get all the resources, including ones created before its registration.


I've added a test that creates some resources and only then registers endpoints, and it doesn't pass - the current code only waits for the endpoints registered at the time it starts to come up.

Patch-set: 4
Reviewer: Gerrit User 618 <618@c72cd27d-1743-4a55-84a4-cd13b8400c52>
Label: Code-Review=-1