VNC replication: Create missing objects on update
When a new replication endpoint is added,
events that happened before that time are not sent to it.

Work around this: make it possible for the user
to manually trigger replication by updating the resources
that haven't been replicated.

When an update fails with "not found",
fetch the resource from GoAPI and try creating it instead.
When a delete fails with "not found", ignore the error.

Log the "not found" errors at Debug level instead of Error.

logging: Use fields instead of putting variables into the message.

Change-Id: Ia2593dd6aa57f1f081847fdfe5de4ebfb9a8a49c
closes-jira-bug: CEM-4349