Support import of SSL enabled contrail cluster
ssl config's are not set in auth proxy client
library, This restricted users to configure
https url's in keystone endpoint.

use SSL_ENABLE and kolla_enable_tls_external key pairs
from contrail_cluster object and openstack_cluster object
to determine whether or not SSL is enabled.

Change-Id: I989b08ca93300d762e7755effce03cc57e84fd20
Closes-Jira-Bug: CEM-7365

Configure https endpoint for analytics

If ssl enabled and the provisioner type is tripleo
analytics endpoint should be configured with https
protocol.

Change-Id: Ida2326050fe6d2ffcff79409fd72779fede7f275
Closes-Jira-Bug: CEM-7934

Add tripleo provisioner type to supported list

add the UT for the triple import case

Change-Id: I23d3f3e7d57ca106117b0dbfc5cda1cdaebf5b21
Closes-Jira-Bug: CEM-7365

Use public port in the endpoint url

when public_api_port is specified in the role objects
it needs to be used in endpoints instead of the default port.

Partial-Jira-Bug: CEM-8208
Partial-Jira-Bug: CEM-7365

Change-Id: Ic2ae9fc51b92e9d45c321ddba0c2d31f48ab33de