make CI faster by not building contrail-base
Currently CI takes 2+ hours just to build a fresh docker image contrail-base.
Switch to last nightly image contrail-base instead.
Drawback is that devs of this repo can no longer Depend-On: changes that
fix/impact contents of contrail-base docker image.

Developer will need to first CI their change in contrail-base, then make it merged,
then wait for the next nightly to succeed for that branch. Only at this point it
would be possible to initiate CI for a change inside Juniper/contrail.

After creating a new branch, CI will not work until a contrail-base is
pushed as tagged with the new branch.

Partial-Jira-Bug: CE-1901
Change-Id: I31b57e3370f867df678617d8ffccaf74c1859666