dynamic proxy: Add X-Service-Token to Swift requests
Long uploads to Swift fail
if the user's token expires during the upload.

Log in as a service user and add X-Service-Token
to requests to endpoints specified in the config.
If Swift is configured to accept service tokens,
it will accept requests with an expired user token
and a valid service token.

dynamicProxyMiddleware:
- Return 400 with no message explicitly.
  echo.NewHTTPError() JSON.Marshal()s the passed error,
  which results in an empty message
  (see https://github.com/golang/go/issues/10748).
  Don't pass the error to echo,
  so that it's obvious that no message is returned to the client.
- Log errors so that the administrator can learn what failed.

Change-Id: I1c366d88aef4056bf63ffd3eeebbd4c7fb5f1d6b
closes-jira-bug: CEM-4676