dynamic proxy: Add X-Service-Token to Swift requests
Long uploads to Swift fail
if the user's token expires during the upload.

Log in as a service user and add X-Service-Token
to requests to endpoints specified in the config.
If Swift is configured to accept service tokens,
it will accept requests with an expired user token
and a valid service token.

If obtaining a service token fails,
skip adding one to the request.

Extract DynamicProxyConfig,
so that only server initialization reads configuration from Viper.

Fix returning error messages to the client.
echo.NewHTTPError() JSON.Marshal()s the passed error,
which results in an empty message
(see https://github.com/golang/go/issues/10748).
Pass a string instead, so that it's marshalled correctly.

Change-Id: I1c366d88aef4056bf63ffd3eeebbd4c7fb5f1d6b
partial-jira-bug: CEM-4676