Detect and report data conversion errors
It was reported in CEM-10012 that a cluster is not brought up with the
proper image using contrail command provisioning cluster cmd.  It was
believed this could be because the API is not giving an error if the
contrail_version is specified as a float (only digits with one decimal).
Add sufficient testing to prove this is the case, and finally to resolve
the issue.

Closes-Jira-Bug: CEM-10291
Change-Id: Ib3689a37f402708de0dcc4c53b62c2214a12c1e7