Change GCP secret generation
Logic for adding path to GCP account to secret.yml was based on existance of provider with type 'gcp' in DB.
In case of deletion of such provider and reprovisioning the cloud the credentials weren't passed to the secret file causing terraform to fail.

Changed the behaviour of GCP secret generation to be the same as AWS and Azure - based on secret file existance instead of provider type.
Removed creation of empty secret files in case none credentials were passed.

Change-Id: I921a1fbf9fff159413f2f111ca8fdeefb75772ea
closes-jira-bug: CEM-10428