Update patch set 12
Patch Set 12:

> Patch Set 12: Code-Review-1
> 
> I would prefer not adding the InterfaceTo*E() variants, but making the existing InterfaceTo*() return errors and changing all the remaining usages to ignore the errors explicitly.
> 
> This way, the functions with the obvious names (InterfaceTo*()) would return errors - new developers would use those. Additionally, the logic that ignores errors would be more readily visible.

I did that at first, and the amount of code changes required was huge.  The InterfaceTo* methods are used in many places.  Debugging the changes in one fell swoop was too much.  I think this approach for now is better.  If we want to get rid of the *E methods in the future we can.

Patch-set: 12
Reviewer: Gerrit User 822 <822@c72cd27d-1743-4a55-84a4-cd13b8400c52>