Remove CloudUser resource
CloudUser resource was used to store credentials for providers (such as AWS and GCP) as well as references to onPrem credentials. After the method of storing provider’s credentials changed usage of this resource was limited to referencing Credential resource.

Goal of this patch is to remove CloudUser and refer Credential directly from Cloud.

Change-Id: I0a2cde8e8868583d7f05f7414b23739e3c0af192
closes-jira-bug: CEM-11000