Update patch set 1
Patch Set 1:

> Patch Set 1: Code-Review-1
> 
> (1 comment)

It's important to understand the cause of issue here:
Azure couldn't connect via IPSec to GCP, because of the GCP firewall blocking UDP traffic.
This meant when doing the healthcheck, it didn't find a tunnel up, marked the container as unhealthy.
Our healthcheck is configured so that after 3 unhealthy checks, the container is restarted, which is what was seen by Vinod in bug report.

This is actually a very important bug, as we don't support IPSec on GCP w/o merging this fix.

Patch-set: 1