Clean modified status of public cloud and fix race
When public cloud is updated check for topology differences is done. If there are differences, update continues and resets modified status of cloud. Otherwise update funcion is left without any modification.

Goal of this patch is to reset modified status in case of identical topologies to prevent situation when it stays modified after successful provisioning.

Additionally this patch fixes the race that occured during extending public topology to OnPrem.
In some cases Cluster package did not wait for the update of existing clouds which caused in broken state.

Change-Id: Iac7005dc9d460f83144d27059265ef7a2b4cbeac
Closes-jira-bug: CEM-11878