Add cluster_id along with deploy_xflow.sh
This will be used in xflow-deploy to get the cluster token id to be passed
to get the node details for this cluster. node details will be used to get
the number of xflow node installed to know to find the diff to delete the
xflow nodes.

Please note that we may not have contrail while xflow deployment, but contrail
command will be there, so in stead of checking config api server for xflow
node details, command db is checked for xflow node count.

related-jira-bug: CEM-9294
Change-Id: I247662e804e005c5ba75ed99c57caac34fa2a4e0