Agent use watcher from sync
Agent module was utilizing polling watcher to monitor changes to cluster information in database.
The polling watcher is not rewrite as the watcher from sync module.
The tasks to run when cluster information changees now resides in agent module, instead of in /etc/contrail/contrail.yml in contrail_command container.

Closes-Jira-Bug: CEM-6496
Change-Id: Iaa908e9dc9d1aa5bad33ca8512af826ed34b23e5